Renamespace JobFactory
